+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2019-06-10
+### Added
+- API Response errors are shown with logger
+- API Response messages
+
+### Fixed
+- NotFound handler fix
+
 ## [1.0.0] - 2019-06-10
 ### Added
 - Project inited